Migrate Support ticket logic to useNotificationsData hook
