+++ packages/manager/src/features/NotificationCenter/NotificationData/index.tsx
@@ -1,0 +1,1 @@
+export { default } from './useNotificationData';
