+++ packages/manager/src/features/NotificationCenter/NotificationData/useNotificationData.tsx
@@ -1,0 +1,11 @@
+import SupportNotifications from './SupportNotifications';
+
+export const useNotificationData = () => {
+  const support = SupportNotifications();
+
+  return {
+    support
+  };
+};
+
+export default useNotificationData;
